Implement "global configuration consistency"
This commit introduces a plugin which adds support for global
configuration consistency. The idea is to try to resolve all
configurations in a way that guarantees that if 2 configurations
have the same dependency, they would use the same version.

There is, in practice, no such guarantee, so in order to make
sure there's no proble, this is implemented using a "strictly"
constraint, which makes sure that the build would fail if the
consistency cannot be reached.

The plugin works by adding an "uberResolve" configuration on
the root project, which includes dependencies from all bucket
configurations of all projects (by default).

This dependency is resolved _before_ any other configuration is
resolved, and then the resolved dependency versions are injected
as constraints to the other resolvable configurations of the
projects.